Add / Refactor code from PoC test
Tavis Ormandy from Google Project Zero reported this Vulnerability to
the Transmission team and wrote up an exploit to demo it. This code is a
slightly refactored version of that code.

See: transmission/transmission#468 for the PR
See: lock.cmpxchg8b.com/Asoquu3e.html for the original exploit page